Rearranging Settings screen with new layout
