+++ src/components/SharedComponents/index.js
@@ -61,6 +61,7 @@ export { default as TextInputSheet } from "./Sheets/TextInputSheet";
 export { default as TextSheet } from "./Sheets/TextSheet";
 export { default as WarningSheet } from "./Sheets/WarningSheet";
 export { default as SimpleObservationLocation } from "./SimpleObservationLocation";
+export { default as SwitchRow } from "./SwitchRow";
 export { default as Tabs } from "./Tabs/Tabs";
 export { default as TaxonResult } from "./TaxonResult";
 export { default as TaxonSearch } from "./TaxonSearch";
