+++ src/i18n/l10n/en.ftl.json
@@ -11,6 +11,7 @@
   "Account-Deleted": "Account Deleted",
   "ACCOUNT-SETTINGS": "ACCOUNT SETTINGS",
   "ACTIVITY": "ACTIVITY",
+  "Activity-Details-View": "Activity & Details View",
   "Add-agreement": "Add agreement",
   "ADD-AN-ID": "ADD AN ID",
   "Add-an-ID-Later": "Add an ID Later",
@@ -29,7 +30,7 @@
   "Add-optional-notes": "Add optional notes",
   "Adds-your-vote-of-agreement": "Adds your vote of agreement",
   "Adds-your-vote-of-disagreement": "Adds your vote of disagreement",
-  "Advanced--interface-mode-with-explainer": "Advanced (Upload multiple photos and sounds)",
+  "ADVANCED-SETTINGS": "ADVANCED SETTINGS",
   "Affiliation": "Affiliation: { $site }",
   "After-capturing-or-importing-photos-show": "After capturing or importing photos, show:",
   "Agree": "Agree",
@@ -39,7 +40,7 @@
   "AI-Camera": "AI Camera",
   "ALL": "ALL",
   "All": "All",
-  "All-observation-options": "All observation options (including iNaturalist AI Camera, Standard Camera, Uploading from Photo Library, and Sound Recorder)",
+  "All-observation-options--list": "All observation options: iNaturalist AI Camera, Standard Camera, Uploading from Photo Library, and Sound Recorder",
   "All-observations": "All observations",
   "All-observations-submitted-to-iNaturalist-need-a-date-and-location": "All observations submitted to iNaturalist need a date and location to be useful to scientists. Please update observations if they need more information.",
   "All-organisms": "All organisms",
@@ -195,7 +196,6 @@
   "datetime-format-short": "M/d/yy h:mm a",
   "datetime-format-short-with-zone": "M/d/yy h:mm a zzz",
   "December": "December",
-  "Default--interface-mode": "Default",
   "DELETE": "DELETE",
   "Delete-all-observations": "Delete all observations",
   "Delete-comment": "Delete comment",
@@ -358,7 +358,6 @@
   "iNaturalist-is-supported-by-our-community": "iNaturalist is supported by our amazing community. From everyday naturalists who add observations and identifications, to curators who manage our taxonomy and help with moderation, to the volunteer translators who make iNaturalist more accessible to worldwide audiences, to our community-based donors, we are extraordinarily grateful to all the people in our community who make iNaturalist the platform it is.",
   "iNaturalist-mission-is-to-connect": "iNaturalist's mission is to connect people to nature and advance biodiversity science and conservation.",
   "INATURALIST-MISSION-VISION": "INATURALIST'S MISSION & VISION",
-  "INATURALIST-MODE": "INATURALIST MODE",
   "INATURALIST-NETWORK": "INATURALIST NETWORK",
   "INATURALIST-SETTINGS": "INATURALIST SETTINGS",
   "INATURALIST-STAFF": "{ $inaturalist } STAFF",
@@ -418,6 +417,7 @@
   "MAP": "MAP",
   "Map-Area": "Map Area",
   "March": "March",
+  "Match-Screen": "Match Screen",
   "maverick--identification": "Maverick",
   "May": "May",
   "MEDIA": "MEDIA",
@@ -491,7 +491,6 @@
   "Obscured": "Obscured",
   "Observation": "Observation",
   "Observation-Attribution": "Observation: © { $userName } · { $restrictions }",
-  "OBSERVATION-BUTTON": "OBSERVATION BUTTON",
   "Observation-Copyright": "Observation Copyright: © { $userName } · { $restrictions }",
   "Observation-has-no-photos-and-no-sounds": "This observation has no photos and no sounds.",
   "Observation-location-obscured-randomized-point": "This observation’s location is obscured. You are seeing a randomized point within the obscuration polygon.",
@@ -735,6 +734,7 @@
   "Sign-in-with-Apple-Failed": "Sign in with Apple Failed",
   "Sign-in-with-Google": "Sign in with Google",
   "Sign-in-with-Google-Failed": "Sign in with Google Failed",
+  "Simple-View": "Simple View",
   "Skip-additional-suggestions": "Skip additional suggestions",
   "Skip-for-now": "Skip for now",
   "Something-went-wrong": "Something went wrong.",
@@ -772,7 +772,6 @@
   "SUBMIT-ID-SUGGESTION": "SUBMIT ID SUGGESTION",
   "SUGGEST-ID": "SUGGEST ID",
   "Suggest-ID": "SUGGEST ID",
-  "SUGGESTIONS": "SUGGESTIONS",
   "supporting--identification": "Supporting",
   "Switches-to-tab": "Switches to { $tab } tab.",
   "Sync-observations": "Sync observations",
@@ -847,6 +846,7 @@
   "Users": "Users",
   "Using-location": "Using location",
   "Version-app-build": "Version { $appVersion } ({ $buildVersion })",
+  "View-Advanced-Settings": "View Advanced Settings",
   "VIEW-ALL-X-PLACES": "VIEW ALL { $count } PLACES",
   "VIEW-ALL-X-PROJECTS": "VIEW ALL { $count } PROJECTS",
   "VIEW-ALL-X-TAXA": "VIEW ALL { $count } TAXA",
@@ -871,6 +871,7 @@
   "WHAT-IS-INATURALIST": "WHAT IS INATURALIST?",
   "Whats-more-by-recording": "What's more, by recording and sharing your observations, you'll create research-quality data for scientists working to better understand and protect nature. So if you like recording your findings from the outdoors, or if you just like learning about life, join us!",
   "When-tapping-the-green-observation-button": "When tapping the green observation button, open:",
+  "When-viewing-observations-display": "When viewing observations, display:",
   "WIKIPEDIA": "WIKIPEDIA",
   "Wild": "Wild",
   "WILD-STATUS": "WILD STATUS",
